Final adjust namespace. Also combine some class into EditLyricRow
