+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Components/Blueprints/TimeTagSelectionBlueprint.cs
@@ -4,11 +4,10 @@
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Game.Rulesets.Edit;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Blueprints
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Blueprints
 {
     public class TimeTagSelectionBlueprint : SelectionBlueprint<TimeTag>
     {
